Update retroarch-003-fix-sdl-screenlock.patch
Incorporates the original fix whilst also fixing this for LF1000 build, stopping the flickering on this platform.  This was being caused by the menu and game video being written to the surface at the same time regardless of which one was active.  Also commented out Window code which is irrelevant to the platform.